pyproject.toml
[tool.poetry]
name = "bananabread"
version = "0.1.0"
description = ""
authors = ["ezzye <ezzy.elliott@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests-html = "^0.10.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
langchain = "^0.0.240"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
